refactor: using path mapping
