+++ src/controllers/UsersController.ts
@@ -1,8 +1,8 @@
-import ConflictException from '../exceptions/ConflictException'
-import UserDuplicated from '../exceptions/UserDuplicated'
-import CreateUserRequest from '../resources/CreateUserRequest'
-import CreateUserResponse from '../resources/CreateUserResponse'
-import register from '../use_cases/register'
+import ConflictException from '@exceptions/ConflictException'
+import UserDuplicated from '@exceptions/UserDuplicated'
+import CreateUserRequest from '@resources/CreateUserRequest'
+import CreateUserResponse from '@resources/CreateUserResponse'
+import register from '@use_cases/register'
 
 const registerController = async (
   request: CreateUserRequest
