+++ src/index.ts
@@ -1,6 +1,6 @@
 import dotenv from 'dotenv'
-import { startServer } from './server'
-import { connect } from './database'
+import { startServer } from '@server/server'
+import { connect } from '@database/database'
 
 const init = async () => {
   dotenv.config()
