+++ src/server/error.ts
@@ -1,8 +1,8 @@
 import express from 'express'
-import ConflictException from '../exceptions/ConflictException'
-import ValidationException from '../exceptions/ValidationException'
-import ErrorResponse from '../resources/ErrorResponse'
-import ValidationErrorResponse from '../resources/ValidationErrorResponse'
+import ConflictException from '@exceptions/ConflictException'
+import ValidationException from '@exceptions/ValidationException'
+import ErrorResponse from '@resources/ErrorResponse'
+import ValidationErrorResponse from '@resources/ValidationErrorResponse'
 
 const respondError = (
   res: express.Response,
